requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
websockets==12.0
python-socketio==5.14.3

aiohttp==3.13.2
python-dotenv==1.2.1
pydantic==2.12.4
python-multipart==0.0.9
python-telegram-bot==20.7

python-dotenv
python-telegram-bot
google-generativeai
agora-token-builder  # <-- ADD THIS
